Changed how the extension works
I changed how the extension works; now it is enable/disable with a
browser button.

I changed the original behaviour of `live-reload` which was exactly the
same script mimosa's live-reload module uses, because the extension
doesn't use `content scripts` any more and it uses `background` ones.

Moreover it also changes the reloading behaviour; it always reload the
page whenever change happens, because I've realised that although it is
a good approach to remap the `css` files, browser doesn't render the
screen if the window doesn't have the focus and the most of the time
that I have the focus on the editor and watching the changes the browser
without swapping the window focus.